+++ src/api/gpx/index.ts
@@ -1,0 +1,7 @@
+export * from "./deleteGpxTrace";
+export * from "./getGpxTrace";
+export * from "./getGpxTraceMetadata";
+export * from "./getOwnGpxTraces";
+export * from "./getTrackPoints";
+export * from "./updateGpxTraceMetadata";
+export * from "./uploadGpxTrace";